wip: c implementation
